move command part out of part.
