+++ mahjong/record/universe/tenhou/event_definition.py
@@ -1,7 +1,8 @@
 from typing import List, Union
 
 from mahjong.record.reader import TenhouGame
-from mahjong.record.universe.format import GameCommand, GameProperty, PlayerView, Update, GameView
+from mahjong.record.universe.format import PlayerView, Update, GameView
+from mahjong.record.universe.command import GameProperty, GameCommand
 from mahjong.record.utils.builder import Builder
 from mahjong.record.utils.event import *
 from mahjong.record.universe.tenhou.xml_macher import tenhou_command
